+++ src/circuits/etable_compact/op_configure/op_bin.rs
@@ -162,7 +162,7 @@ impl<F: FieldExt> EventTableOpcodeConfig<F> for BinConfig {
             _ => unreachable!(),
         };
 
-        self.vtype.assign(ctx, vtype as u16)?;
+        self.vtype.assign(ctx, vtype as u64)?;
         self.lhs.assign(ctx, left)?;
         self.rhs.assign(ctx, right)?;
         self.res.assign(ctx, value)?;
